pyproject.toml
[project]
name = "autosign"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arabert>=1.0.1",
    "mediapipe==0.10.9",
    "natsort>=8.4.0",
    "pandas>=3.0.0",
    "pillow==10.4.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "transformers==4.45.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0",
    "python-dotenv>=1.0.0",
]
